Исправил бажок с онли y = 1 (раньше откадывало дробную часть на 1/5)

+++ src/com/company/Zvd1.java
@@ -19,7 +19,7 @@ public class Zvd1 {
             y = Math.pow(x, 2);
         }
         else {
-            y = Math.pow(Math.exp(Math.cos(x)), 1 / 5);
+            y = Math.pow(Math.exp(Math.cos(x)), ((double)1 / (double)5));
         }
     }
